+++ test/baselines/adapter_TS_React/admin/src/components/settings.tsx
@@ -85,7 +85,7 @@ class Settings extends React.Component<SettingsProps, SettingsState> {
 		return (
 			<FormControl
 				className={`${this.props.classes.input} ${this.props.classes.controlElement}`}
-				style={{ 
+				style={{
 					paddingTop: 5,
 					...style
 				}}
@@ -110,7 +110,7 @@ class Settings extends React.Component<SettingsProps, SettingsState> {
 		return (
 			<FormControlLabel
 				key={attr}
-				style={{ 
+				style={{
 					paddingTop: 5,
 					...style
 				}}